requirements.txt
discord.py
luhn